Añado fichero de test #24
Muevo los tests a un archivo para el "comprobador"
